fix: remove old run_grid_search script (closes #31)
Replace the old standalone run_grid_search.py with the refactored version that properly integrates
with the sampleworks package.
