+++ run_grid_search.py
@@ -3,61 +3,24 @@
 from __future__ import annotations
 
 import argparse
+import concurrent
 import csv
 import json
-import logging
+import multiprocessing
 import os
+import pickle
 import shutil
 import subprocess
 import time
-from concurrent.futures import ProcessPoolExecutor
+from concurrent.futures import ProcessPoolExecutor  # pyright: ignore
 from dataclasses import asdict, dataclass
-from datetime import datetime
 from pathlib import Path
-from queue import Queue
 from threading import Lock
+from typing import Any
 
-
-logging.basicConfig(
-    level=logging.INFO,
-    format="[%(asctime)s] %(message)s",
-    datefmt="%Y-%m-%d %H:%M:%S",
-)
-log = logging.getLogger(__name__)
-
-
-@dataclass
-class JobConfig:
-    protein: str
-    structure_path: str
-    density_path: str
-    resolution: float
-    model: str
-    scaler: str
-    ensemble_size: int
-    gradient_weight: float
-    gd_steps: int
-    method: str | None
-    output_dir: str
-    log_path: str
-
-
-@dataclass
-class JobResult:
-    protein: str
-    model: str
-    method: str | None
-    scaler: str
-    ensemble_size: int
-    gradient_weight: float
-    gd_steps: int
-    status: str
-    exit_code: int
-    runtime_seconds: float
-    started_at: str
-    finished_at: str
-    log_path: str
-    output_dir: str
+from loguru import logger as log
+from sampleworks.utils import guidance_constants as constants
+from sampleworks.utils.guidance_script_arguments import GuidanceConfig, JobConfig, JobResult
 
 
 @dataclass
@@ -121,115 +84,201 @@ def detect_gpus() -> list[str]:
 
 
 def get_pixi_env(model: str) -> str:
-    if model in ("boltz1", "boltz2"):
-        return "boltz"
-    elif model == "protenix":
-        return "protenix"
-    elif model == "rf3":
-        return "rf3"
-    else:
-        raise ValueError(f"Unknown model: {model}")
+    match model:
+        case constants.BOLTZ_1 | constants.BOLTZ_2:
+            return "boltz"
+        case constants.PROTENIX | constants.RF3:
+            return model
+        case _:
+            raise ValueError(
+                f"Unknown model: {model}. Valid options are: {constants.BOLTZ_1}, "
+                f"{constants.BOLTZ_2}, {constants.PROTENIX}, {constants.RF3}"
+            )
 
 
-def get_checkpoint(model: str, args: argparse.Namespace) -> str | None:
-    if model == "boltz1":
-        return args.boltz1_checkpoint
-    elif model == "boltz2":
-        return args.boltz2_checkpoint
-    elif model == "protenix":
-        return args.protenix_checkpoint
-    elif model == "rf3":
-        return args.rf3_checkpoint
-    return None
-
-
-def build_command(job: JobConfig, args: argparse.Namespace) -> list[str]:
-    env = get_pixi_env(job.model)
-    script_dir = Path(__file__).parent
-    script_name = f"{job.model}_{job.scaler}.py"
-    script_path = script_dir / "scripts" / script_name
-
-    cmd = ["pixi", "run", "-e", env, "python", str(script_path)]
-
-    cmd.extend(["--ensemble-size", str(job.ensemble_size)])
-    cmd.extend(["--structure", job.structure_path])
-    cmd.extend(["--density", job.density_path])
-    cmd.extend(["--resolution", str(job.resolution)])
-    cmd.extend(["--output-dir", job.output_dir])
-    cmd.extend(["--partial-diffusion-step", str(args.partial_diffusion_step)])
-    cmd.extend(["--loss-order", str(args.loss_order)])
-    cmd.extend(["--device", "cuda:0"])
-
-    checkpoint = get_checkpoint(job.model, args)
-    if checkpoint:
-        cmd.extend(["--model-checkpoint", checkpoint])
-
-    if job.model == "boltz2" and job.method:
-        cmd.extend(["--method", job.method])
-
-    if job.scaler == "fk_steering":
-        cmd.extend(["--guidance-weight", str(job.gradient_weight)])
-        cmd.extend(["--num-gd-steps", str(job.gd_steps)])
-        cmd.extend(["--num-particles", str(args.num_particles)])
-        cmd.extend(["--fk-lambda", str(args.fk_lambda)])
-        cmd.extend(["--fk-resampling-interval", str(args.fk_resampling_interval)])
-    else:
-        cmd.extend(["--step-size", str(job.gradient_weight)])
-        if args.use_tweedie:
-            cmd.append("--use-tweedie")
-
-    if args.gradient_normalization:
-        cmd.append("--gradient-normalization")
-    if args.augmentation:
-        cmd.append("--augmentation")
-    if args.align_to_input:
-        cmd.append("--align-to-input")
-
-    return cmd
-
-
-def run_job(
-    job: JobConfig, gpu: str, args: argparse.Namespace, clean_output: bool = False
-) -> JobResult:
-    started_at = datetime.now().isoformat()
-    start_time = time.time()
-
-    if clean_output and os.path.exists(job.output_dir):
-        log.info(f"Cleaning existing output directory: {job.output_dir}")
-        shutil.rmtree(job.output_dir)
-
-    os.makedirs(os.path.dirname(job.log_path), exist_ok=True)
-
-    cmd = build_command(job, args)
-    env = os.environ.copy()
-    env["CUDA_VISIBLE_DEVICES"] = gpu
-
-    log.info(f"Starting on GPU {gpu}: {job.log_path}")
-
-    with open(job.log_path, "w") as log_file:
-        result = subprocess.run(cmd, stdout=log_file, stderr=subprocess.STDOUT, env=env)
-
-    runtime = time.time() - start_time
-    finished_at = datetime.now().isoformat()
-
-    status = "success" if result.returncode == 0 else "failed"
-
-    return JobResult(
+def build_args_for_process_pool(
+    job: JobConfig, args: argparse.Namespace, device_num: int | None = None
+) -> GuidanceConfig:
+    guidance_config = GuidanceConfig(
         protein=job.protein,
+        structure=job.structure_path,
+        density=job.density_path,
         model=job.model,
-        method=job.method,
-        scaler=job.scaler,
-        ensemble_size=job.ensemble_size,
-        gradient_weight=job.gradient_weight,
-        gd_steps=job.gd_steps,
-        status=status,
-        exit_code=result.returncode,
-        runtime_seconds=round(runtime, 2),
-        started_at=started_at,
-        finished_at=finished_at,
+        guidance_type=job.scaler,
         log_path=job.log_path,
         output_dir=job.output_dir,
+        loss_order=args.loss_order,
+        partial_diffusion_step=args.partial_diffusion_step,
+        resolution=job.resolution,
+        device=f"cuda:{device_num}" if device_num is not None else "",
+        gradient_normalization=args.gradient_normalization,
+        augmentation=args.augmentation,
+        align_to_input=args.align_to_input,
     )
+    # given model_type and guidance_type, the GuidanceConfig class will set itself up
+    # with defaults for remaining required args, but we want to set them further here.
+    guidance_config.populate_config_for_guidance_type(job, args)
+    return guidance_config
+
+
+def run_grid_search(
+    jobs: list[JobConfig],
+    gpus: list[str],
+    args: argparse.Namespace,
+    job_statuses: dict[int, str] | None = None,
+) -> list[JobResult]:
+    """
+    Replacing run_job and run_grid_search, avoiding model reloads
+    of the model.
+    Args:
+        jobs: generated by generate_and_filter_jobs, a list of JobConfig objects
+        gpus: the available GPUs to run jobs on, really we only use the length of the list
+        args: command-line arguments to this script, used to pass some on to jobs.
+        job_statuses: a dictionary mapping job IDs to their statuses,
+            also generated by generate_and_filter_jobs.
+
+    Returns:
+        list of JobResult objects, one for each job run, primarily to track success/failure
+    """
+    results: list[JobResult] = []
+    successful = 0
+    failed = 0
+
+    max_workers = len(gpus)
+    log.info(f"Running {len(jobs)} jobs with {max_workers} parallel workers")
+
+    # Divide the job among the workers:
+    worker_job_queues = [
+        [build_args_for_process_pool(j, args, i) for j in jobs[i::max_workers]]
+        for i in range(max_workers)
+    ]
+    # we'll pickle each job queue separately and then execute each job queue in a separate process
+    # in principle we could just pass the job queue directly to the worker_wrapper function, but
+    # this keeps a record of what we did, which may be useful for debugging.
+    job_queue_paths = []
+    for wjq in worker_job_queues:
+        wjq_path = os.path.join(args.output_dir, f"wjq_{id(wjq)}.pkl")
+        log.info(f"Pickling worker job queue to {wjq_path}")
+        job_queue_paths.append(wjq_path)
+        with open(wjq_path, "wb") as f:
+            pickle.dump(wjq, f)
+
+    if args.dry_run:
+        log.info(f"[DRY-RUN] Running {len(jobs)} jobs with {max_workers} parallel workers")
+        log.info(f"[DRY-RUN] Job queue paths: {job_queue_paths}")
+        return results
+
+    # Clean up output directories if they already exist and have failed previously:
+    for i, job in enumerate(jobs):
+        clean_output = False
+        if job_statuses is not None:
+            clean_output = job_statuses.get(id(job), "not_run") != "not_run"
+        if clean_output and os.path.exists(job.output_dir):
+            log.info(f"Cleaning existing output directory: {job.output_dir}")
+            shutil.rmtree(job.output_dir)
+
+    # TODO this approach works, but I think it may be more efficient to actually call a
+    #  script that runs the jobs (a script that basically only calls run_guidance_job_queue
+    #  to avoid pickling objects.
+    with ProcessPoolExecutor(
+        max_workers=max_workers,  # TODO: may need to tune this or make a flag.
+        mp_context=multiprocessing.get_context("spawn"),
+    ) as executor:
+        futures = {}
+
+        for worker_num, job_queue_path in enumerate(job_queue_paths):
+            model = worker_job_queues[worker_num][0].model
+            future = executor.submit(
+                run_guidance_queue_script, (job_queue_path, max_workers, model, worker_num)
+            )
+            futures[future] = job_queue_path
+
+        for completed in concurrent.futures.as_completed(futures):  # pyright: ignore
+            try:
+                with open(futures[completed].replace(".pkl", ".results.pkl"), "rb") as f:
+                    result = pickle.load(f)
+
+                results.extend(result)
+                for r in result:
+                    if r.status == "success":
+                        successful += 1
+                        log.info(
+                            f"SUCCESS ({r.protein}, {r.model}, {r.method}, {r.scaler} "
+                            f"{r.runtime_seconds:.1f}s): {r.log_path}"
+                        )
+                    else:
+                        failed += 1
+                        log.error(
+                            f"FAILED ({r.protein}, {r.model}, {r.method}, {r.scaler} "
+                            f"exit={r.exit_code}): {r.log_path}"
+                        )
+            except Exception as e:
+                failed += 1
+                log.error(f"Job failed with exception: {e}")  # this won't be very informative
+    return results
+
+
+def run_guidance_queue_script(args: tuple[str, int, str, int]):
+    job_queue_path, max_workers, model, worker_num = args
+    pixi_env = get_pixi_env(model)
+    script_path = Path(__file__).parent / "examples" / "run_guidance_pipeline.py"
+    cmd = f"pixi run -e {pixi_env} python {script_path} --job-queue-path {job_queue_path}"
+    cmd = cmd.split()
+    log.info(f"Running worker {worker_num}: {cmd} on GPU {worker_num % max_workers}")
+    # env = os.environ.copy()
+
+    with open(job_queue_path.replace(".pkl", ".log"), "w") as log_file:
+        result = subprocess.run(cmd, stdout=log_file, stderr=subprocess.STDOUT)
+    return result
+
+
+def main(args: argparse.Namespace):
+    """
+    Main pipeline for running grid search experiments.
+    Args:
+        args: Command-line arguments.
+    """
+    gpus = detect_gpus()
+    log.info(f"Detected {len(gpus)} GPUs: {gpus}")
+    if args.max_parallel != "auto":
+        gpus = gpus[: int(args.max_parallel)]
+
+    log_args(args, gpus)
+
+    if len(args.models.split()) > 1:
+        # this is designed to run one type of model per script, # TODO to allow multiple models
+        raise ValueError("Multiple --models selected, this is not compatible with the new script!")
+    if len(args.methods.split(",")) > 1:
+        # this is designed to run one type of model per script, # TODO to allow multiple models
+        raise ValueError("Multiple --methods selected, this is not compatible with the new script!")
+
+    filtered_jobs, job_statuses = generate_and_filter_jobs(args)
+
+    if len(filtered_jobs) == 0:
+        log.info("No jobs to run!")
+        return
+
+    config = GridSearchConfig(
+        models=args.models.split(),
+        scalers=args.scalers.split(),
+        ensemble_sizes=[int(x) for x in args.ensemble_sizes.split()],
+        gradient_weights=[float(x) for x in args.gradient_weights.split()],
+        gd_steps=[int(x) for x in args.num_gd_steps.split()],
+        methods=[m.strip() for m in args.methods.split(",")],
+        proteins_file=args.proteins,
+        output_dir=args.output_dir,
+    )
+
+    start_time = time.time()
+
+    results = run_grid_search(filtered_jobs, gpus, args, job_statuses=job_statuses)
+
+    if not args.dry_run and results:
+        save_results(results, config, args.output_dir, time.time() - start_time)
+
+    log.info("=" * 50)
+    log.info("Grid search complete")
+    log.info("=" * 50)
 
 
 def generate_jobs(args: argparse.Namespace) -> list[JobConfig]:
@@ -319,89 +368,6 @@ def generate_jobs(args: argparse.Namespace) -> list[JobConfig]:
 
 
 print_lock = Lock()
-
-
-def worker_wrapper(args_tuple: tuple) -> JobResult:
-    job, gpu, args, clean_output = args_tuple
-    return run_job(job, gpu, args, clean_output)
-
-
-def run_grid_search(
-    jobs: list[JobConfig],
-    gpus: list[str],
-    args: argparse.Namespace,
-    job_statuses: dict[int, str] | None = None,
-) -> list[JobResult]:
-    results: list[JobResult] = []
-    successful = 0
-    failed = 0
-
-    if args.dry_run:
-        for job in jobs:
-            cmd = build_command(job, args)
-            log.info(f"[DRY-RUN] {' '.join(cmd)}")
-        return results
-
-    gpu_queue: Queue[str] = Queue()
-    for gpu in gpus:
-        gpu_queue.put(gpu)
-
-    max_workers = len(gpus)
-    log.info(f"Running {len(jobs)} jobs with {max_workers} parallel workers")
-
-    with ProcessPoolExecutor(max_workers=max_workers) as executor:
-        futures = {}
-        pending_jobs = list(jobs)
-        active_futures = set()
-
-        while pending_jobs or active_futures:
-            while pending_jobs and not gpu_queue.empty():
-                job = pending_jobs.pop(0)
-                gpu = gpu_queue.get()
-
-                clean_output = False
-                if job_statuses is not None:
-                    job_status = job_statuses.get(id(job), "not_run")
-                    clean_output = job_status != "not_run"
-
-                future = executor.submit(worker_wrapper, (job, gpu, args, clean_output))
-                futures[future] = gpu
-                active_futures.add(future)
-
-            if active_futures:
-                done_futures = set()
-                for future in active_futures:
-                    if future.done():
-                        done_futures.add(future)
-
-                for future in done_futures:
-                    active_futures.remove(future)
-                    gpu = futures[future]
-                    gpu_queue.put(gpu)
-
-                    try:
-                        result = future.result()
-                        results.append(result)
-                        if result.status == "success":
-                            successful += 1
-                            log.info(
-                                f"SUCCESS (GPU {gpu}, {result.runtime_seconds:.1f}s): "
-                                f"{result.log_path}"
-                            )
-                        else:
-                            failed += 1
-                            log.info(
-                                f"FAILED (GPU {gpu}, exit={result.exit_code}): {result.log_path}"
-                            )
-                    except Exception as e:
-                        failed += 1
-                        log.error(f"Job failed with exception: {e}")
-
-                if not done_futures and active_futures:
-                    time.sleep(1)
-
-    log.info(f"Completed: {successful} successful, {failed} failed")
-    return results
 
 
 def save_results(
@@ -567,13 +533,7 @@ def parse_args() -> argparse.Namespace:
     return parser.parse_args()
 
 
-def main():
-    args = parse_args()
-
-    gpus = detect_gpus()
-    if args.max_parallel != "auto":
-        gpus = gpus[: int(args.max_parallel)]
-
+def log_args(args: argparse.Namespace, gpus: list[str]):
     log.info("=" * 50)
     log.info("Starting grid search")
     log.info(f"Models: {args.models}")
@@ -587,6 +547,10 @@ def main():
     log.info(f"Dry run: {args.dry_run}")
     log.info("=" * 50)
 
+
+# TODO make job statuses a proper class
+# TODO: there are many constants here like "not_run" that should be defined in only one place.
+def generate_and_filter_jobs(args: argparse.Namespace) -> tuple[list[JobConfig], dict[Any, Any]]:
     jobs = generate_jobs(args)
     log.info(f"Generated {len(jobs)} total jobs")
 
@@ -616,33 +580,9 @@ def main():
     else:
         filtered_jobs = [job for job in jobs if job_statuses[id(job)] in ("failed", "not_run")]
         log.info(f"Running failed and un-run jobs (default): {len(filtered_jobs)} jobs")
-
-    if len(filtered_jobs) == 0:
-        log.info("No jobs to run!")
-        return
-
-    config = GridSearchConfig(
-        models=args.models.split(),
-        scalers=args.scalers.split(),
-        ensemble_sizes=[int(x) for x in args.ensemble_sizes.split()],
-        gradient_weights=[float(x) for x in args.gradient_weights.split()],
-        gd_steps=[int(x) for x in args.num_gd_steps.split()],
-        methods=[m.strip() for m in args.methods.split(",")],
-        proteins_file=args.proteins,
-        output_dir=args.output_dir,
-    )
-
-    start_time = time.time()
-    results = run_grid_search(filtered_jobs, gpus, args, job_statuses)
-    total_time = time.time() - start_time
-
-    if not args.dry_run and results:
-        save_results(results, config, args.output_dir, total_time)
-
-    log.info("=" * 50)
-    log.info("Grid search complete")
-    log.info("=" * 50)
+    return filtered_jobs, job_statuses
 
 
 if __name__ == "__main__":
-    main()
+    args = parse_args()
+    main(args)
